Use scoped enumerations in Network module

+++ examples/ftp/Ftp.cpp
@@ -15,7 +15,7 @@
 ////////////////////////////////////////////////////////////
 std::ostream& operator<<(std::ostream& stream, const sf::Ftp::Response& response)
 {
-    return stream << response.getStatus() << response.getMessage();
+    return stream << static_cast<int>(response.getStatus()) << response.getMessage();
 }
